pyproject.toml
[project]
name = "integrify-docs"
version = "1.0.0"
description = "Integrify üçün dokumentasiya repositoriyası."
authors = [
    { name = "mmzeynalli", email = "miradil.zeynalli@gmail.com" },
    { name = "vahidzhe", email = "vahidzhe@gmail.com" },
]
requires-python = "~=3.12"
readme = "README.md"
license = "GPL-3.0-or-later"
dependencies = [
    "pydantic>=2.11.7",
    "mkdocs-material[imaging]>=9.6.4,<10",
    "mkdocstrings[python]>=0.28.1,<1.0.0",
    "mkdocs-panzoom-plugin>=0.1.3,<0.2",
    "griffe-pydantic>=1.1.0,<2",
]

[project.urls]
Homepage = "https://integrify.mmzeynalli.dev/"
Repository = "https://github.com/Integrify-SDK/integrify-docs-python"
Documentation = "https://integrify.mmzeynalli.dev/"

[project.entry-points."mkdocs.plugins"]
generate_nav = "plugins.generate_nav:IntegrationsNavPlugin"
